globalized login and made use of it in ProfileHeader
