+++ instatonne-frontend/src/app/components/login/login.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, AfterViewInit, ChangeDetectorRef, ApplicationRef, Ng
 import { NgForm } from '@angular/forms';
 import { AuthService } from 'src/app/services/auth.service';
 import { Observable, of, Subject } from 'rxjs';
-import { tap } from 'rxjs/operators';
+import { filter } from 'rxjs/operators';
 import { Router } from '@angular/router';
 import { User } from 'src/app/generated/models';
 
@@ -19,47 +19,19 @@ export class LoginComponent implements OnInit {
 
   constructor(
     private authService: AuthService,
-    private ngZone: NgZone,
     private router: Router
   ) { }
 
   ngOnInit(): void {
-    this.loadGAPI();
-    window['onSignIn'] = this.onSignIn.bind(this);
-
     this.user$ = this.authService.getUser();
     this.authUser$ = this.authService.getAuthUser();
-  }
 
-  public onSignIn(googleUser: gapi.auth2.GoogleUser) {
-    const profile = googleUser.getBasicProfile();
-    console.log('ID: ' + profile.getId());
-    console.log('Name: ' + profile.getName());
-    console.log('Image URL: ' + profile.getImageUrl());
-    console.log('Email: ' + profile.getEmail());
-    console.log(googleUser.getAuthResponse().id_token);
-    this.ngZone.run(() => {
-      this.authService.setAuthUser(googleUser);
+    this.user$.pipe(filter(user => !!user)).subscribe(_ => {
+      this.router.navigate(['profile']);
     });
   }
 
-  private loadGAPI() {
-    let isFound = false;
-    const scripts = Array.from(document.getElementsByTagName('script'));
-    for (const script of scripts) {
-      if (script.getAttribute('src') != null && script.getAttribute('src').includes('https://apis.google.com/js/platform.js')) {
-        isFound = true;
-      }
-    }
 
-    if (!isFound) {
-      const node = document.createElement('script');
-      node.src = 'https://apis.google.com/js/platform.js';
-      node.type = 'text/javascript';
-      node.async = false;
-      document.getElementsByTagName('head')[0].appendChild(node);
-    }
-  }
 
   onRegister(f: NgForm) {
     if (!f.valid) { return; }
